[clang-tidy] Add bsl-dependent-base-name
Add a check that warns on potentially unexpected non-dependent name
bindings resolved in a class template. Specifically, emit a warning
whenever a non-dependent, unqualified name is used in a class template
with at least one dependent base, and that name is declared in 1) an
enclosing scope of the template and 2) at least one of its dependent
bases.

The purpose of this check is to ensure that the names that are bound
in a template match what the developer expects.

Note that names that reference type parameters in explicit template
instantiations are not checked. This is due to the apparent inability
to retrieve the name of a template type _as_written_ in the source
(i.e. before it is desuarged in the AST) while simultaneously
knowing the template argument's position relative to its enclosing
nodes (like a class or function) in the AST.